% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_all_reservoirs}
\alias{calibrate_all_reservoirs}
\title{calibrate_all_reservoirs}
\usage{
calibrate_all_reservoirs(data_path, output = "parameters")
}
\arguments{
\item{data_path}{path to data directory "ERCOT Reservoir Watershed Delineations and Inflow Scenarios"}

\item{output}{return either optimized parameters ("parameters") or simulation results ("results")}
}
\description{
calibrate_all_reservoirs
}
\details{
...
}
