% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_texas_reservoirs}
\alias{plot_texas_reservoirs}
\title{plot_texas_watershed}
\usage{
plot_texas_reservoirs(drawdown_scenario)
}
\arguments{
\item{drawdown_scenario}{tibble with reservoir name and storage}
}
\description{
plot_texas_watershed
}
\details{
plot ERCOT reservoir points by Summer Capacity and Storage Level
}
