% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_watershed}
\alias{plot_watershed}
\title{plot_watershed}
\usage{
plot_watershed(reservoir)
}
\arguments{
\item{reservoir}{name of reservoir (or reservoirs as vector)}
}
\description{
plot_watershed
}
\details{
plot watershed for a named reservoir or set of reservoirs
}
