% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_controlflows}
\alias{read_controlflows}
\title{read_controlflows}
\usage{
read_controlflows(reservoir, data_path)
}
\arguments{
\item{reservoir}{name of reservoir}

\item{data_path}{path to data directory "ERCOT Reservoir Watershed Delineations and Inflow Scenarios"}
}
\description{
read_controlflows
}
\details{
...
}
