% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_ERCOT_data}
\alias{read_ERCOT_data}
\title{read_ERCOT_data}
\usage{
read_ERCOT_data(res_filepath)
}
\arguments{
\item{res_filepath}{full path to reservoir data file for ERCOT}
}
\description{
read_ERCOT_data
}
\details{
...
}
