% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{simulate_gcm_inflows_all_dams}
\alias{simulate_gcm_inflows_all_dams}
\title{simulate_gcm_inflows_all_dams}
\usage{
simulate_gcm_inflows_all_dams(res_filepath, period)
}
\arguments{
\item{res_filepath}{full path to reservoir data file for ERCOT}

\item{period}{either "baseline" or "future"}
}
\description{
simulate_gcm_inflows_all_dams
}
\details{
...
}
