% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{validate_all_reservoirs}
\alias{validate_all_reservoirs}
\title{validate_all_reservoirs}
\usage{
validate_all_reservoirs(res_filepath)
}
\arguments{
\item{res_filepath}{full path to reservoir records and attributes file ("ERCOT_reservoir_attributes_and_records_PUBLIC.xlsx")}
}
\description{
validate_all_reservoirs
}
\details{
perform k-fold validation on all reservoir models
}
