% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_all_reservoirs}
\alias{calibrate_all_reservoirs}
\title{calibrate_all_reservoirs}
\usage{
calibrate_all_reservoirs(res_filepath)
}
\arguments{
\item{res_filepath}{full path to reservoir data file for ERCOT}
}
\description{
calibrate_all_reservoirs
}
\details{
...
}
