% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_reservoir_model}
\alias{calibrate_reservoir_model}
\title{calibrate_reservoir_model}
\usage{
calibrate_reservoir_model(
  reservoir_name,
  res_filepath,
  plot = FALSE,
  s_cap_def = "reported",
  output = "parameters",
  validation_mode = FALSE
)
}
\arguments{
\item{reservoir_name}{name of reservoir to be calibrated}

\item{res_filepath}{full path to reservoir records and attributes file ("ERCOT_reservoir_attributes_and_records_PUBLIC.xlsx")}

\item{plot}{T/F plot storage time series?}

\item{output}{return either optimized parameters ("parameters") or simulation results ("results")}

\item{validation_mode}{logical. Set to TRUE to perform k-fold validation on model and report RMSE scores across validation groups.}
}
\description{
calibrate_reservoir_model
}
\details{
Performs calibration of reservoir model using monthly demand factors
}
