% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{generate_monthly_flow_replicates}
\alias{generate_monthly_flow_replicates}
\title{generate_monthly_flow_replicates}
\usage{
generate_monthly_flow_replicates(reservoir)
}
\arguments{
\item{reservoir}{name of the reservoir}
}
\description{
generate_monthly_flow_replicates
}
\details{
...
}
