% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_controlflows}
\alias{read_controlflows}
\title{read_controlflows}
\usage{
read_controlflows(reservoir)
}
\arguments{
\item{res_filepath}{full path to reservoir data file for ERCOT}
}
\description{
read_controlflows
}
\details{
...
}
