% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_gcm_flows}
\alias{read_gcm_flows}
\title{read_gcm_flows}
\usage{
read_gcm_flows(reservoir, gcm, period)
}
\arguments{
\item{reservoir}{...}

\item{gcm}{gcm}

\item{period}{"baseline" or "future"}
}
\description{
read_gcm_flows
}
\details{
...
}
