% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{simulate_gcm_inflows}
\alias{simulate_gcm_inflows}
\title{simulate_gcm_inflows}
\usage{
simulate_gcm_inflows(
  res_filepath,
  reservoir,
  period,
  parameters,
  s_initial = 1
)
}
\arguments{
\item{res_filepath}{full path to reservoir data file for ERCOT}

\item{reservoir}{name of the reservoir}

\item{period}{either "baseline" or "future"}

\item{parameters}{tibble with headers of "target", "hedge", and "reservoir"}

\item{output}{return either optimized parameters ("parameters") or simulation results ("results")}
}
\description{
simulate_gcm_inflows
}
\details{
...
}
