% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{simulate_gcm_inflows_all_dams}
\alias{simulate_gcm_inflows_all_dams}
\title{simulate_gcm_inflows_all_dams}
\usage{
simulate_gcm_inflows_all_dams(data_path, period, params_path)
}
\arguments{
\item{data_path}{path to data directory "ERCOT Reservoir Watershed Delineations and Inflow Scenarios"}

\item{period}{either "baseline" or "future"}
}
\description{
simulate_gcm_inflows_all_dams
}
\details{
...
}
