% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{validate_all_reservoirs}
\alias{validate_all_reservoirs}
\title{validate_all_reservoirs}
\usage{
validate_all_reservoirs(data_path)
}
\arguments{
\item{data_path}{path to data directory "ERCOT Reservoir Watershed Delineations and Inflow Scenarios"}
}
\description{
validate_all_reservoirs
}
\details{
perform k-fold validation on all reservoir models
}
